refactor: migrate and rename functions, classes and files; fix small errors
